+++ tests/tsconfig.json
@@ -20,7 +20,7 @@
     "lib": ["dom", "es2019"],
     "types": ["jest", "node"]
   },
-  "include": ["./src/**/*"],
+  "include": ["system/**/*"],
   "exclude": ["node_modules"],
   "atom": {
     "rewriteTsconfig": false
